+++ src/kernel/emulation/linux/psynch/psynch_cvwait.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 long sys_psynch_cvwait(void* cv, uint32_t cvgen, uint32_t cvugen, void* mutex, uint32_t mgen,
-		uint32_t ugen, uint64_t sec, uint64_t usec);
+		uint32_t ugen, uint64_t sec, uint32_t usec);
 
 #endif
